+++ src/android/ContactsX.java
@@ -1,15 +1,24 @@
 package de.einfachhans.ContactsX;
 
 import android.Manifest;
+import android.accounts.Account;
+import android.accounts.AccountManager;
 import android.app.Activity;
+import android.content.ContentProviderOperation;
+import android.content.ContentProviderResult;
 import android.content.ContentResolver;
+import android.content.ContentValues;
 import android.content.Intent;
+import android.content.OperationApplicationException;
 import android.database.Cursor;
+import android.net.Uri;
+import android.os.RemoteException;
 import android.provider.ContactsContract;
 
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CallbackContext;
 
+import org.apache.cordova.LOG;
 import org.apache.cordova.PermissionHelper;
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -26,10 +35,14 @@ import java.util.Map;
 public class ContactsX extends CordovaPlugin {
 
     private CallbackContext _callbackContext;
+    private final String LOG_TAG = "ContactsX";
 
     public static final String READ = Manifest.permission.READ_CONTACTS;
+    public static final String WRITE = Manifest.permission.WRITE_CONTACTS;
 
-    public static final int REQ_CODE_READ = 0;
+    private static final String EMAIL_REGEXP = ".+@.+\\.+.+"; /* <anything>@<anything>.<anything>*/
+
+    public static final int REQ_CODE_PERMISSIONS = 0;
     public static final int REQ_CODE_PICK = 2;
 
     @Override
@@ -43,16 +56,29 @@ public class ContactsX extends CordovaPlugin {
                 } else {
                     returnError(ContactsXErrorCodes.PermissionDenied);
                 }
-            } else if(action.equals("pick")) {
+            } else if (action.equals("pick")) {
                 if (PermissionHelper.hasPermission(this, READ)) {
-                    this.pick(args);
+                    this.pick();
+                } else {
+                    returnError(ContactsXErrorCodes.PermissionDenied);
+                }
+            } else if (action.equals("save")) {
+                if(PermissionHelper.hasPermission(this, WRITE)) {
+                    this.save(args);
+                } else {
+                    returnError(ContactsXErrorCodes.PermissionDenied);
+                }
+            } else if(action.equals("delete")) {
+                if(PermissionHelper.hasPermission(this, WRITE)) {
+                    this.delete(args);
                 } else {
                     returnError(ContactsXErrorCodes.PermissionDenied);
                 }
             } else if (action.equals("hasPermission")) {
                 this.hasPermission();
             } else if (action.equals("requestPermission")) {
-                this.requestPermission(REQ_CODE_READ);
+                boolean write = args.optBoolean(0);
+                this.requestPermission(write);
             } else {
                 returnError(ContactsXErrorCodes.UnsupportedAction);
             }
@@ -66,11 +92,11 @@ public class ContactsX extends CordovaPlugin {
     }
 
     public void onActivityResult(int requestCode, int resultCode, final Intent intent) {
-        if(requestCode == REQ_CODE_PICK) {
-            if(resultCode == Activity.RESULT_OK) {
+        if (requestCode == REQ_CODE_PICK) {
+            if (resultCode == Activity.RESULT_OK) {
                 String contactId = intent.getData().getLastPathSegment();
-                Cursor c =  this.cordova.getActivity().getContentResolver().query(ContactsContract.RawContacts.CONTENT_URI,
-                        new String[] {ContactsContract.RawContacts._ID}, ContactsContract.RawContacts.CONTACT_ID + " = " + contactId, null, null);
+                Cursor c = this.cordova.getActivity().getContentResolver().query(ContactsContract.RawContacts.CONTENT_URI,
+                        new String[]{ContactsContract.RawContacts._ID}, ContactsContract.RawContacts.CONTACT_ID + " = " + contactId, null, null);
                 if (!c.moveToFirst()) {
                     returnError(ContactsXErrorCodes.UnknownError, "Error occurred while retrieving contact raw id");
                     return;
@@ -79,7 +105,7 @@ public class ContactsX extends CordovaPlugin {
                 c.close();
 
                 JSONObject contact = getContactById(id);
-                if(contact != null) {
+                if (contact != null) {
                     this._callbackContext.success(contact);
                 } else {
                     returnError(ContactsXErrorCodes.UnknownError);
@@ -139,6 +165,7 @@ public class ContactsX extends CordovaPlugin {
         projection.add(ContactsContract.Data.MIMETYPE);
         projection.add(ContactsContract.Contacts._ID);
         projection.add(ContactsContract.Data.CONTACT_ID);
+        projection.add(ContactsContract.Data.RAW_CONTACT_ID);
         projection.add(ContactsContract.CommonDataKinds.Contactables.DATA);
 
         if (options.displayName) {
@@ -153,7 +180,7 @@ public class ContactsX extends CordovaPlugin {
         if (options.familyName) {
             projection.add(ContactsContract.CommonDataKinds.StructuredName.FAMILY_NAME);
         }
-        if(options.emails) {
+        if (options.emails) {
             projection.add(ContactsContract.CommonDataKinds.Email._ID);
             projection.add(ContactsContract.CommonDataKinds.Email.DATA);
             projection.add(ContactsContract.CommonDataKinds.Email.TYPE);
@@ -189,6 +216,9 @@ public class ContactsX extends CordovaPlugin {
                 String contactId = contactsCursor.getString(
                         contactsCursor.getColumnIndex(ContactsContract.Data.CONTACT_ID)
                 );
+                String rawId = contactsCursor.getString(
+                        contactsCursor.getColumnIndex(ContactsContract.Data.RAW_CONTACT_ID)
+                );
 
                 JSONObject jsContact = new JSONObject();
 
@@ -197,6 +227,7 @@ public class ContactsX extends CordovaPlugin {
                     // so put it to the HashMap
 
                     jsContact.put("id", contactId);
+                    jsContact.put("rawId", rawId);
                     if (options.displayName) {
                         String displayName = contactsCursor.getString(contactsCursor.getColumnIndex(ContactsContract.Contacts.DISPLAY_NAME));
                         jsContact.put("displayName", displayName);
@@ -215,15 +246,12 @@ public class ContactsX extends CordovaPlugin {
                 String mimeType = contactsCursor.getString(
                         contactsCursor.getColumnIndex(ContactsContract.Data.MIMETYPE)
                 );
-                String data = contactsCursor.getString(
-                        contactsCursor.getColumnIndex(ContactsContract.CommonDataKinds.Contactables.DATA)
-                );
 
                 assert jsContact != null;
                 switch (mimeType) {
                     case ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE:
                         JSONArray jsPhoneNumbers = jsContact.getJSONArray("phoneNumbers");
-                        jsPhoneNumbers.put(data);
+                        jsPhoneNumbers.put(phoneQuery(contactsCursor));
                         break;
                     case ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE:
                         JSONArray emailAddresses = jsContact.getJSONArray("emails");
@@ -257,6 +285,17 @@ public class ContactsX extends CordovaPlugin {
         return jsContacts;
     }
 
+    private JSONObject phoneQuery(Cursor cursor) throws JSONException {
+        JSONObject phoneNumber = new JSONObject();
+        int typeCode = cursor.getInt(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.TYPE));
+        String typeLabel = cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.LABEL));
+        String type = (typeCode == ContactsContract.CommonDataKinds.Phone.TYPE_CUSTOM) ? typeLabel : getPhoneType(typeCode);
+        phoneNumber.put("id", cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone._ID)));
+        phoneNumber.put("value", cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.NUMBER)));
+        phoneNumber.put("type", type);
+        return phoneNumber;
+    }
+
     private JSONObject emailQuery(Cursor cursor) throws JSONException {
         JSONObject email = new JSONObject();
         int typeCode = cursor.getInt(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Email.TYPE));
@@ -279,14 +318,11 @@ public class ContactsX extends CordovaPlugin {
 
             if ("home".equals(lowerType)) {
                 return ContactsContract.CommonDataKinds.Email.TYPE_HOME;
-            }
-            else if ("work".equals(lowerType)) {
+            } else if ("work".equals(lowerType)) {
                 return ContactsContract.CommonDataKinds.Email.TYPE_WORK;
-            }
-            else if ("other".equals(lowerType)) {
+            } else if ("other".equals(lowerType)) {
                 return ContactsContract.CommonDataKinds.Email.TYPE_OTHER;
-            }
-            else if ("mobile".equals(lowerType)) {
+            } else if ("mobile".equals(lowerType)) {
                 return ContactsContract.CommonDataKinds.Email.TYPE_MOBILE;
             }
             return ContactsContract.CommonDataKinds.Email.TYPE_CUSTOM;
@@ -320,7 +356,7 @@ public class ContactsX extends CordovaPlugin {
         return stringType;
     }
 
-    private void pick(JSONArray args) {
+    private void pick() {
         this.cordova.getThreadPool().execute(() -> {
             Intent contactPickerIntent = new Intent(Intent.ACTION_PICK, ContactsContract.Contacts.CONTENT_URI);
             this.cordova.startActivityForResult(this, contactPickerIntent, REQ_CODE_PICK);
@@ -343,7 +379,7 @@ public class ContactsX extends CordovaPlugin {
 
         try {
             JSONArray contacts = handleFindResult(c, new ContactsXFindOptions(new JSONObject(pickFields)));
-            if(contacts.length() == 1) {
+            if (contacts.length() == 1) {
                 return contacts.getJSONObject(0);
             }
         } catch (Exception e) {
@@ -353,16 +389,351 @@ public class ContactsX extends CordovaPlugin {
         return null;
     }
 
+    private void save(JSONArray args) throws JSONException {
+        final JSONObject contact = args.getJSONObject(0);
+        this.cordova.getThreadPool().execute(() -> {
+            JSONObject res = null;
+            String id = performSave(contact);
+            if (id != null) {
+                res = getContactById(id);
+            }
+            if (res != null) {
+                _callbackContext.success(res);
+            } else {
+                returnError(ContactsXErrorCodes.UnknownError);
+            }
+        });
+    }
+
+    private String performSave(JSONObject contact) {
+        AccountManager mgr = AccountManager.get(this.cordova.getActivity());
+        Account[] accounts = mgr.getAccounts();
+        String accountName = null;
+        String accountType = null;
+
+        if (accounts.length == 1) {
+            accountName = accounts[0].name;
+            accountType = accounts[0].type;
+        } else if (accounts.length > 1) {
+            for (Account a : accounts) {
+                if (a.type.contains("eas") && a.name.matches(EMAIL_REGEXP)) /*Exchange ActiveSync*/ {
+                    accountName = a.name;
+                    accountType = a.type;
+                    break;
+                }
+            }
+            if (accountName == null) {
+                for (Account a : accounts) {
+                    if (a.type.contains("com.google") && a.name.matches(EMAIL_REGEXP)) /*Google sync provider*/ {
+                        accountName = a.name;
+                        accountType = a.type;
+                        break;
+                    }
+                }
+            }
+            if (accountName == null) {
+                for (Account a : accounts) {
+                    if (a.name.matches(EMAIL_REGEXP)) /*Last resort, just look for an email address...*/ {
+                        accountName = a.name;
+                        accountType = a.type;
+                        break;
+                    }
+                }
+            }
+        }
+
+        String id = getJsonString(contact, "id");
+        if (id == null) {
+            // Create new contact
+            return newContact(contact, accountType, accountName);
+        } else {
+            // Modify existing contact
+            return modifyContact(id, contact, accountType, accountName);
+        }
+    }
+
+    private String newContact(JSONObject contact, String accountType, String accountName) {
+        // Create a list of attributes to add to the contact database
+        ArrayList<ContentProviderOperation> ops = new ArrayList<ContentProviderOperation>();
+
+        //Add contact type
+        ops.add(ContentProviderOperation.newInsert(ContactsContract.RawContacts.CONTENT_URI)
+                .withValue(ContactsContract.RawContacts.ACCOUNT_TYPE, accountType)
+                .withValue(ContactsContract.RawContacts.ACCOUNT_NAME, accountName)
+                .build());
+
+        // Add name
+        String displayName = getJsonString(contact, "displayName");
+        String firstName = getJsonString(contact, "firstName");
+        String middleName = getJsonString(contact, "middleName");
+        String familyName = getJsonString(contact, "familyName");
+        if (displayName != null || firstName != null || middleName != null || familyName != null) {
+            ops.add(ContentProviderOperation.newInsert(ContactsContract.Data.CONTENT_URI)
+                    .withValueBackReference(ContactsContract.Data.RAW_CONTACT_ID, 0)
+                    .withValue(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.StructuredName.CONTENT_ITEM_TYPE)
+                    .withValue(ContactsContract.CommonDataKinds.StructuredName.DISPLAY_NAME, displayName)
+                    .withValue(ContactsContract.CommonDataKinds.StructuredName.FAMILY_NAME, familyName)
+                    .withValue(ContactsContract.CommonDataKinds.StructuredName.MIDDLE_NAME, middleName)
+                    .withValue(ContactsContract.CommonDataKinds.StructuredName.GIVEN_NAME, firstName)
+                    .build());
+        } else {
+            LOG.d(LOG_TAG, "All \"name\" properties are empty");
+        }
+
+        //Add phone numbers
+        JSONArray phones;
+        try {
+            phones = contact.getJSONArray("phoneNumbers");
+            for (int i = 0; i < phones.length(); i++) {
+                if (!phones.isNull(i)) {
+                    JSONObject phone = (JSONObject) phones.get(i);
+                    ops.add(ContentProviderOperation.newInsert(ContactsContract.Data.CONTENT_URI)
+                            .withValueBackReference(ContactsContract.Data.RAW_CONTACT_ID, 0)
+                            .withValue(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE)
+                            .withValue(ContactsContract.CommonDataKinds.Phone.NUMBER, getJsonString(phone, "value"))
+                            .withValue(ContactsContract.CommonDataKinds.Phone.TYPE, getPhoneType(getJsonString(phone, "type")))
+                            .withValue(ContactsContract.CommonDataKinds.Phone.LABEL, getJsonString(phone, "type"))
+                            .build());
+                }
+            }
+        } catch (JSONException e) {
+            LOG.d(LOG_TAG, "Could not get phone numbers");
+        }
+
+        // Add emails
+        JSONArray emails;
+        try {
+            emails = contact.getJSONArray("emails");
+            for (int i = 0; i < emails.length(); i++) {
+                JSONObject email = (JSONObject) emails.get(i);
+                ops.add(ContentProviderOperation.newInsert(ContactsContract.Data.CONTENT_URI)
+                        .withValueBackReference(ContactsContract.Data.RAW_CONTACT_ID, 0)
+                        .withValue(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE)
+                        .withValue(ContactsContract.CommonDataKinds.Email.DATA, getJsonString(email, "value"))
+                        .withValue(ContactsContract.CommonDataKinds.Email.TYPE, getContactType(getJsonString(email, "type")))
+                        .withValue(ContactsContract.CommonDataKinds.Email.LABEL, getJsonString(email, "type"))
+                        .build());
+            }
+        } catch (JSONException e) {
+            LOG.d(LOG_TAG, "Could not get emails");
+        }
+
+        String newId = null;
+        //Add contact
+        try {
+            ContentProviderResult[] cpResults = this.cordova.getActivity().getContentResolver().applyBatch(ContactsContract.AUTHORITY, ops);
+            if (cpResults.length > 0) {
+                newId = cpResults[0].uri.getLastPathSegment();
+            }
+        } catch (RemoteException | OperationApplicationException e) {
+            LOG.e(LOG_TAG, e.getMessage(), e);
+        }
+        return newId;
+    }
+
+    private String modifyContact(String id, JSONObject contact, String accountType, String accountName) {
+        // Get the RAW_CONTACT_ID which is needed to insert new values in an already existing contact.
+        // But not needed to update existing values.
+        String rawId = getJsonString(contact, "rawId");
+
+        // Create a list of attributes to add to the contact database
+        ArrayList<ContentProviderOperation> ops = new ArrayList<ContentProviderOperation>();
+
+        //Add contact type
+        ops.add(ContentProviderOperation.newUpdate(ContactsContract.RawContacts.CONTENT_URI)
+                .withValue(ContactsContract.RawContacts.ACCOUNT_TYPE, accountType)
+                .withValue(ContactsContract.RawContacts.ACCOUNT_NAME, accountName)
+                .build());
+
+        // Modify name
+        String displayName = getJsonString(contact, "displayName");
+        String firstName = getJsonString(contact, "firstName");
+        String middleName = getJsonString(contact, "middleName");
+        String familyName = getJsonString(contact, "familyName");
+        if (displayName != null || firstName != null || middleName != null || familyName != null) {
+            ContentProviderOperation.Builder builder = ContentProviderOperation.newUpdate(ContactsContract.Data.CONTENT_URI)
+                    .withSelection(ContactsContract.Data.CONTACT_ID + "=? AND " +
+                                    ContactsContract.Data.MIMETYPE + "=?",
+                            new String[]{id, ContactsContract.CommonDataKinds.StructuredName.CONTENT_ITEM_TYPE});
+
+            if (displayName != null) {
+                builder.withValue(ContactsContract.CommonDataKinds.StructuredName.DISPLAY_NAME, displayName);
+            }
+
+            if (familyName != null) {
+                builder.withValue(ContactsContract.CommonDataKinds.StructuredName.FAMILY_NAME, familyName);
+            }
+            if (middleName != null) {
+                builder.withValue(ContactsContract.CommonDataKinds.StructuredName.MIDDLE_NAME, middleName);
+            }
+            if (firstName != null) {
+                builder.withValue(ContactsContract.CommonDataKinds.StructuredName.GIVEN_NAME, firstName);
+            }
+
+            ops.add(builder.build());
+        }
+
+        // Modify phone numbers
+        JSONArray phones;
+        try {
+            phones = contact.getJSONArray("phoneNumbers");
+            // Delete all the phones
+            if (phones.length() == 0) {
+                ops.add(ContentProviderOperation.newDelete(ContactsContract.Data.CONTENT_URI)
+                        .withSelection(ContactsContract.Data.RAW_CONTACT_ID + "=? AND " +
+                                        ContactsContract.Data.MIMETYPE + "=?",
+                                new String[]{"" + rawId, ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE})
+                        .build());
+            }
+            // Modify or add a phone
+            else {
+                for (int i = 0; i < phones.length(); i++) {
+                    JSONObject phone = (JSONObject) phones.get(i);
+                    String phoneId = getJsonString(phone, "id");
+                    // This is a new phone so do a DB insert
+                    if (phoneId == null) {
+                        ContentValues contentValues = new ContentValues();
+                        contentValues.put(ContactsContract.Data.RAW_CONTACT_ID, rawId);
+                        contentValues.put(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE);
+                        contentValues.put(ContactsContract.CommonDataKinds.Phone.NUMBER, getJsonString(phone, "value"));
+                        contentValues.put(ContactsContract.CommonDataKinds.Phone.TYPE, getPhoneType(getJsonString(phone, "type")));
+                        contentValues.put(ContactsContract.CommonDataKinds.Phone.LABEL, getJsonString(phone, "type"));
+
+                        ops.add(ContentProviderOperation.newInsert(
+                                ContactsContract.Data.CONTENT_URI).withValues(contentValues).build());
+                    }
+                    // This is an existing phone so do a DB update
+                    else {
+                        ops.add(ContentProviderOperation.newUpdate(ContactsContract.Data.CONTENT_URI)
+                                .withSelection(ContactsContract.CommonDataKinds.Phone._ID + "=? AND " +
+                                                ContactsContract.Data.MIMETYPE + "=?",
+                                        new String[]{phoneId, ContactsContract.CommonDataKinds.Phone.CONTENT_ITEM_TYPE})
+                                .withValue(ContactsContract.CommonDataKinds.Phone.NUMBER, getJsonString(phone, "value"))
+                                .withValue(ContactsContract.CommonDataKinds.Phone.TYPE, getPhoneType(getJsonString(phone, "type")))
+                                .withValue(ContactsContract.CommonDataKinds.Phone.LABEL, getJsonString(phone, "type"))
+                                .build());
+                    }
+                }
+            }
+        } catch (JSONException e) {
+            LOG.d(LOG_TAG, "Could not get phone numbers");
+        }
+
+        // Modify emails
+        JSONArray emails;
+        try {
+            emails = contact.getJSONArray("emails");
+            // Delete all the emails
+            if (emails.length() == 0) {
+                ops.add(ContentProviderOperation.newDelete(ContactsContract.Data.CONTENT_URI)
+                        .withSelection(ContactsContract.Data.RAW_CONTACT_ID + "=? AND " +
+                                        ContactsContract.Data.MIMETYPE + "=?",
+                                new String[]{"" + rawId, ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE})
+                        .build());
+            }
+            // Modify or add a email
+            else {
+                for (int i = 0; i < emails.length(); i++) {
+                    JSONObject email = (JSONObject) emails.get(i);
+                    String emailId = getJsonString(email, "id");
+                    // This is a new email so do a DB insert
+                    if (emailId == null) {
+                        ContentValues contentValues = new ContentValues();
+                        contentValues.put(ContactsContract.Data.RAW_CONTACT_ID, rawId);
+                        contentValues.put(ContactsContract.Data.MIMETYPE, ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE);
+                        contentValues.put(ContactsContract.CommonDataKinds.Email.DATA, getJsonString(email, "value"));
+                        contentValues.put(ContactsContract.CommonDataKinds.Email.TYPE, getContactType(getJsonString(email, "type")));
+                        contentValues.put(ContactsContract.CommonDataKinds.Email.LABEL, getJsonString(email, "type"));
+
+                        ops.add(ContentProviderOperation.newInsert(
+                                ContactsContract.Data.CONTENT_URI).withValues(contentValues).build());
+                    }
+                    // This is an existing email so do a DB update
+                    else {
+                        String emailValue = getJsonString(email, "value");
+                        if (!emailValue.isEmpty()) {
+                            ops.add(ContentProviderOperation.newUpdate(ContactsContract.Data.CONTENT_URI)
+                                    .withSelection(ContactsContract.CommonDataKinds.Email._ID + "=? AND " +
+                                                    ContactsContract.Data.MIMETYPE + "=?",
+                                            new String[]{emailId, ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE})
+                                    .withValue(ContactsContract.CommonDataKinds.Email.DATA, getJsonString(email, "value"))
+                                    .withValue(ContactsContract.CommonDataKinds.Email.TYPE, getContactType(getJsonString(email, "type")))
+                                    .withValue(ContactsContract.CommonDataKinds.Email.LABEL, getJsonString(email, "type"))
+                                    .build());
+                        } else {
+                            ops.add(ContentProviderOperation.newDelete(ContactsContract.Data.CONTENT_URI)
+                                    .withSelection(ContactsContract.CommonDataKinds.Email._ID + "=? AND " +
+                                                    ContactsContract.Data.MIMETYPE + "=?",
+                                            new String[]{emailId, ContactsContract.CommonDataKinds.Email.CONTENT_ITEM_TYPE})
+                                    .build());
+                        }
+                    }
+                }
+            }
+        } catch (JSONException e) {
+            LOG.d(LOG_TAG, "Could not get emails");
+        }
+
+        boolean retVal = true;
+        //Modify contact
+        try {
+            this.cordova.getActivity().getContentResolver().applyBatch(ContactsContract.AUTHORITY, ops);
+        } catch (RemoteException | OperationApplicationException e) {
+            LOG.e(LOG_TAG, e.getMessage(), e);
+            retVal = false;
+        }
+
+        // if the save was a success return the contact ID
+        if (retVal) {
+            return rawId;
+        } else {
+            return null;
+        }
+    }
+
+    private void delete(JSONArray args) throws JSONException {
+        final String contactId = args.getString(0);
+        this.cordova.getThreadPool().execute(() -> {
+            if (performDelete(contactId)) {
+                _callbackContext.success();
+            } else {
+                returnError(ContactsXErrorCodes.UnknownError);
+            }
+        });
+    }
+
+    private boolean performDelete(String id) {
+        int result = 0;
+        Cursor cursor = this.cordova.getActivity().getContentResolver().query(ContactsContract.Contacts.CONTENT_URI,
+                null,
+                ContactsContract.Contacts._ID + " = ?",
+                new String[] { id }, null);
+
+        if (cursor.getCount() == 1) {
+            cursor.moveToFirst();
+            String lookupKey = cursor.getString(cursor.getColumnIndex(ContactsContract.Contacts.LOOKUP_KEY));
+            Uri uri = Uri.withAppendedPath(ContactsContract.Contacts.CONTENT_LOOKUP_URI, lookupKey);
+            result = this.cordova.getActivity().getContentResolver().delete(uri, null, null);
+        } else {
+            LOG.d(LOG_TAG, "Could not find contact with ID");
+        }
+
+        cursor.close();
+
+        return result > 0;
+    }
+
     private void hasPermission() throws JSONException {
         JSONObject response = new JSONObject();
         response.put("read", PermissionHelper.hasPermission(this, READ));
+        response.put("write", PermissionHelper.hasPermission(this, WRITE));
         if (this._callbackContext != null) {
             this._callbackContext.success(response);
         }
     }
 
-    private void requestPermission(int requestCode) {
-        PermissionHelper.requestPermission(this, requestCode, READ);
+    private void requestPermission(boolean write) {
+        PermissionHelper.requestPermission(this, REQ_CODE_PERMISSIONS, write ? WRITE : READ);
     }
 
     private void returnError(ContactsXErrorCodes errorCode) {
@@ -377,5 +748,155 @@ public class ContactsX extends CordovaPlugin {
             _callbackContext.error(new JSONObject(resultMap));
             _callbackContext = null;
         }
+    }
+
+    // Helper
+
+    private String getJsonString(JSONObject obj, String property) {
+        String value = null;
+        try {
+            if (obj != null) {
+                value = obj.getString(property);
+            }
+        } catch (JSONException e) {
+            LOG.d(LOG_TAG, "Could not get = " + e.getMessage());
+        }
+        return value;
+    }
+
+    /**
+     * getPhoneType converts an Android phone type into a string
+     *
+     * @param type
+     * @return phone type as string.
+     */
+    private String getPhoneType(int type) {
+        String stringType;
+
+        switch (type) {
+            case ContactsContract.CommonDataKinds.Phone.TYPE_CUSTOM:
+                stringType = "custom";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_FAX_HOME:
+                stringType = "home fax";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_FAX_WORK:
+                stringType = "work fax";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_HOME:
+                stringType = "home";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_MOBILE:
+                stringType = "mobile";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_PAGER:
+                stringType = "pager";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_WORK:
+                stringType = "work";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_CALLBACK:
+                stringType = "callback";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_CAR:
+                stringType = "car";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_COMPANY_MAIN:
+                stringType = "company main";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_OTHER_FAX:
+                stringType = "other fax";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_RADIO:
+                stringType = "radio";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_TELEX:
+                stringType = "telex";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_TTY_TDD:
+                stringType = "tty tdd";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_WORK_MOBILE:
+                stringType = "work mobile";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_WORK_PAGER:
+                stringType = "work pager";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_ASSISTANT:
+                stringType = "assistant";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_MMS:
+                stringType = "mms";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_ISDN:
+                stringType = "isdn";
+                break;
+            case ContactsContract.CommonDataKinds.Phone.TYPE_OTHER:
+            default:
+                stringType = "other";
+                break;
+        }
+        return stringType;
+    }
+
+    /**
+     * Converts a string from the W3C Contact API to it's Android int value.
+     *
+     * @param string
+     * @return Android int value
+     */
+    private int getPhoneType(String string) {
+
+        int type = ContactsContract.CommonDataKinds.Phone.TYPE_OTHER;
+
+        if (string != null) {
+            String lowerType = string.toLowerCase(Locale.getDefault());
+
+            switch (lowerType) {
+                case "home":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_HOME;
+                case "mobile":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_MOBILE;
+                case "work":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_WORK;
+                case "work fax":
+                case "fax":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_FAX_WORK;
+                case "home fax":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_FAX_HOME;
+                case "pager":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_PAGER;
+                case "other":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_OTHER;
+                case "car":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_CAR;
+                case "company main":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_COMPANY_MAIN;
+                case "isdn":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_ISDN;
+                case "main":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_MAIN;
+                case "other fax":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_OTHER_FAX;
+                case "radio":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_RADIO;
+                case "telex":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_TELEX;
+                case "work mobile":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_WORK_MOBILE;
+                case "work pager":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_WORK_PAGER;
+                case "assistant":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_ASSISTANT;
+                case "mms":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_MMS;
+                case "callback":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_CALLBACK;
+                case "tty ttd":
+                    return ContactsContract.CommonDataKinds.Phone.TYPE_TTY_TDD;
+            }
+            return ContactsContract.CommonDataKinds.Phone.TYPE_CUSTOM;
+        }
+        return type;
     }
 }